add message passing logic to separate file
