+++ hippynn/custom_kernels/__init__.py
@@ -47,6 +47,7 @@ from typing import Union
 import torch
 from .registry import MessagePassingKernels
 from . import env_pytorch
+from .message_passing_torch import torchMessagePassingHop, torchMessagePassingQuad, torchMessagePassingVec
 from .. import settings
 
 class CustomKernelError(Exception):
@@ -218,7 +219,7 @@ del try_custom_kernels
 # If they will not import, then the kernel active will not end up being triton, so
 # this will not create an issue.
 try:
-    from .tensor_message_passing import tensorMessagePassingHop, tensorMessagePassingBackwardOnly, tensorMessagePassingVec, tensorMessagePassingQuad
+    from .message_passing_triton import tensorMessagePassingHop, tensorMessagePassingBackwardOnly, tensorMessagePassingVec, tensorMessagePassingQuad
 except:
     pass
 
@@ -230,42 +231,17 @@ def hopMessagePassing(T,s,z,pair_first,pair_second):
         else:
             return tensorMessagePassingBackwardOnly(T,s,z,pair_first,pair_second)
     else:
-        ij,t = T.shape
-        _,nu = s.shape
-        sensitivity = s.unsqueeze(1) * T.unsqueeze(2)
-        sense_flat = sensitivity.reshape(ij, t * nu)
-        return envsum(sense_flat, z, pair_first, pair_second)
+        return torchMessagePassingHop(T,s,z,pair_first,pair_second,envsum)
     
 # Wrapper for tensor message passing for HIP-NN-VEC
 def vecMessagePassing(in_features, sense_vals, pair_first, pair_second, dist_pairs, coord_pairs):
     if in_features.device != 'cpu' and kernel_active == 'triton' and settings.USE_ENV_TENSOR_GRADIENT:
         return tensorMessagePassingVec(in_features, sense_vals, pair_first, pair_second, dist_pairs, coord_pairs)
     else:
-        _, nu = sense_vals.shape
-        sense_vec = sense_vals.unsqueeze(1) * (coord_pairs / dist_pairs.unsqueeze(1)).unsqueeze(2)
-        sense_vec = sense_vec.reshape(-1, nu * 3)
-        sense_stacked = torch.concatenate([sense_vals, sense_vec], dim=1)
-        return envsum(sense_stacked, in_features, pair_first, pair_second)
+        return torchMessagePassingVec(in_features, sense_vals, pair_first, pair_second, dist_pairs, coord_pairs, envsum)
     
 def quadMessagePassing(in_features, sense_vals, pair_first, pair_second, dist_pairs, coord_pairs, upper_ind):
     if in_features.device != 'cpu' and kernel_active == 'triton' and settings.USE_ENV_TENSOR_GRADIENT:
         return tensorMessagePassingQuad(in_features, sense_vals, pair_first, pair_second, dist_pairs, coord_pairs)
     else:
-        _, nu = sense_vals.shape
-
-        rhats = coord_pairs / dist_pairs.unsqueeze(1)
-        sense_vec = sense_vals.unsqueeze(1) * rhats.unsqueeze(2)
-        sense_vec = sense_vec.reshape(-1, nu * 3)
-        rhatsquad = rhats.unsqueeze(1) * rhats.unsqueeze(2)
-        rhatsquad = (rhatsquad + rhatsquad.transpose(1, 2)) / 2
-        tr = torch.diagonal(rhatsquad, dim1=1, dim2=2).sum(dim=1) / 3.0  # Add divide by 3 early to save flops
-        tr = tr.unsqueeze(1).unsqueeze(2) * torch.eye(3, dtype=tr.dtype, device=tr.device).unsqueeze(0)
-        rhatsquad = rhatsquad - tr
-        rhatsqflat = rhatsquad.reshape(-1, 9)[:, upper_ind]  # Upper-diagonal part
-        sense_quad = sense_vals.unsqueeze(1) * rhatsqflat.unsqueeze(2)
-        sense_quad = sense_quad.reshape(-1, nu * 5)
-        sense_stacked = torch.concatenate([sense_vals, sense_vec, sense_quad], dim=1)
-
-        # Message passing, stack sensitivities to coalesce custom kernel call.
-        # shape (n_atoms, n_nu + 3*n_nu + 5*n_nu, n_feat)
-        return envsum(sense_stacked, in_features, pair_first, pair_second)
+        return torchMessagePassingQuad(in_features, sense_vals, pair_first, pair_second, dist_pairs, coord_pairs, upper_ind, envsum)
